requirements.txt
pyyaml
selenium
Appium-Python-Client